消息

+++ dictationhelper/app/src/main/java/cn/edu/hebtu/software/listendemo/Host/listenWord/ListenResultSelectActivity.java
@@ -190,7 +190,7 @@ public class ListenResultSelectActivity extends AppCompatActivity {
                     AlertDialog.Builder adBuilder = new AlertDialog.Builder(ListenResultSelectActivity.this);
                     adBuilder.setCancelable(false);
                     adBuilder.setTitle("     本次听写成绩：");
-                    adBuilder.setMessage("            " + score * 100 + "分");
+                    adBuilder.setMessage("            " + (int)(score * 100) + "分");
                     adBuilder.setPositiveButton("返回首页", new DialogInterface.OnClickListener() {
                         @Override
                         public void onClick(DialogInterface dialog, int which) {
